backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.12.0
huggingface_hub>=0.20.0
Pillow>=9.0.0
numpy>=1.23.0
opencv-python-headless>=4.7.0
matplotlib>=3.7.0
